Have FirebaseArray accessible from subclasses
Problem:
Currently, I think, It's not possible to remove items from the FirebaseListAdapter without removing it also on firebase db. 

Proposal:
I think (not sure) that having access directly to FirebaseArray would allow to create in the subclass custom adapter a method to remove items **only** on the adpater and not in db.