Update README.md
I've seen (and answered some) a lot of question in StackOverflow that complains about how their RecyclerView not get updated. And some of them got it fixed after they only change the recyclerView.setHasFixedSize to false. And I wonder why they make it true in the first place if they don't know what it does, so I come here and check, and there it is. I don't know why it is there, so please decline my change if it does any kind of harm to this great repository.