Provided scope for twitter and facebook sdk
Problem:
Apps are currently required to add the facebook and twitter dependency
even if they do not require these login methods to be supported.

Solution:
In this change, we assume a provided scope dependency on facebook and
twitter sdks and expect that the apps provide these dependencies.
If these dependencies  are not found, we fail early with an actionable
error.
Also, the size of the resulting apk reduces by about 2MB.

Change-Id: I5979039c281dd6daca265f60f9c9dcf59596ced1