Fixing race between user code submission and verification callback
Problems:
We run into a potential race when the device is rotated and the confirmation code is submitted before
the phone verification callbacks reinstate the verification id and related params.

Solution:
We preserve the verification id and related state and reinstate it before the callbacks overwrite it.
In addition, we add speculative checks to make sure that verification id and confirmation code are not null.

Change-Id: Ia7245e3dd17995560480577a5adcdb87bb681ba3